requirements.txt
pyyaml>=6.0.1
pyqt5>=5.15.9
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=14.0.0
openpyxl>=3.1.2
